nvim: start migrating to lazy structure
Moving the plugin structure to that recommended by `lazy.nvim`. Lazy
loading `nvim-tree` and `telescope` on keys. This has the added benefit
of defining the keys where the plugin is defined rather than a central
`mappings.lua`.

Right now `mappings.lua` is not being loaded but I've kept it around
to remind me which other keys I need to migrate to the new structure.